added Ahmeds latest code, change the names and reduced the kf* files to 1
